Recognition/requirements.txt
torch==1.7.1
torchvision==0.8.2
tqdm==4.42.1
pillow==7.0.0
numpy==1.18.1
